requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
google-cloud-vision==3.7.0
gspread==6.1.0
google-auth==2.29.0
